add boilerplate for the frontend
