+++ .gitignore
@@ -1,1 +1,3 @@
 console-for-sap
+webapp/frontend/assets
+webapp/frontend/node_modules
