+++ cmd/root.go
@@ -8,6 +8,8 @@ import (
 
 	"github.com/mitchellh/go-homedir"
 	"github.com/spf13/viper"
+
+	"github.com/SUSE/console-for-sap/cmd/webapp"
 )
 
 var cfgFile string
@@ -29,6 +31,7 @@ func Execute() {
 func init() {
 	cobra.OnInitialize(initConfig)
 	rootCmd.PersistentFlags().StringVar(&cfgFile, "config", "", "config file (default is $HOME/.console-for-sap.yaml)")
+	rootCmd.AddCommand(webapp.NewWebappCmd())
 }
 
 // initConfig reads in config file and ENV variables if set.
