+++ webapp/engine.go
@@ -3,19 +3,25 @@ package webapp
 import (
 	"embed"
 	"html/template"
+	"net/http"
 
 	"github.com/gin-gonic/gin"
 )
 
 //go:embed templates
-var FS embed.FS
+var templatesFS embed.FS
 
-func Engine() *gin.Engine {
+//go:embed frontend/assets
+var assetsFS embed.FS
 
-	templates := template.Must(template.New("").ParseFS(FS, "templates/*.tmpl"))
+func NewEngine() *gin.Engine {
+
+	templates := template.Must(template.New("").ParseFS(templatesFS, "templates/*.tmpl"))
 
 	engine := gin.Default()
 	engine.SetHTMLTemplate(templates)
+
+	engine.StaticFS("/static", http.FS(assetsFS))
 	engine.GET("/", homeHandler)
 
 	return engine
