+++ webapp/frontend/stylesheets/stylesheets.scss
@@ -1,0 +1,3 @@
+@import "../node_modules/eos-ds/dist/vendors/bootstrap/css/bootstrap";
+@import "../node_modules/eos-ds/dist/scss/index";
+
