+++ webapp/templates/home.html.tmpl
@@ -1,9 +1,12 @@
 <!doctype html>
-<html>
+<html lang="en">
   <head>
     <title>{{ .title }}</title>
+    <link rel="stylesheet" type="text/css" href="static/frontend/assets/stylesheets.css" />
   </head>
   <body>
-    <h1>{{ .title }}</h1>
+    <div class="container-fluid">
+      <header class="col"><h1>{{ .title }}</h1></header>
+    </div>
   </body>
 </html>
